requirements.txt
google-genai
loguru
beautifulsoup4
requests
